+++ v2/app.py
@@ -3,4 +3,4 @@ from flask import Flask
 app = Flask(__name__)
 
 if __name__ == "__main__":
-    app.run(debug=False, threaded=False, use_reloader=False)
+    app.run(debug=False, threaded=False, use_reloader=True)